Differentiate admins and all CAS users
This allows anyone to log into the USG site, but normal users won't see
anything different.

There is currently no way to make a user an admin through the web
interface, but something like
  User.all.map{|u| u.update_attribute(:is_admin, true) }
should ensure the same functionality.

This adds a controller before_filter :authenticate_admin! which will
ensure the user is an admin, as well as a helper current_user_is_admin?
to mimic user_signed_in?, but for admins.