Stick the footer to the bottom of the browser
Use the technique here:
  http://ryanfait.com/resources/footer-stick-to-bottom-of-page/
to attach the footer to the bottom of the browser.

Basically, this puts all non-footer body content in a div with a
negative margin-bottom equal to the height of the footer.

This alone is enough to position the footer always at the bottom of the
page; but when the body content is too long it overlaps the footer, this
step is not enough. For this, a "push" div is added to make sure there
is at least #footer's height of blank space at the bottom of <body> for
the footer to overlap.